Split tests in many files to improve paralelization
